+++ R/gggenes-package.R
@@ -2,5 +2,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom lifecycle deprecate_warn
 ## usethis namespace: end
 NULL