+++ tests/testthat/test-feature.R
@@ -3,7 +3,8 @@ library(ggplot2)
 test_that("geom_feature() and geom_feature_label() in Cartesian coordinates", {
   p <- base_cartesian() + 
          geom_feature(data = example_features, 
-                     aes(x = position, y = molecule, forward = forward)) +
+                     aes(x = position, y = molecule, forward = forward),
+                     linewidth = 1) +
          geom_feature_label(data = example_features, 
                            aes(x = position, y = molecule, label = name, 
                                forward = forward)) + 
@@ -18,7 +19,8 @@ test_that("geom_feature() and geom_feature_label() in Cartesian coordinates", {
 test_that("geom_feature() and geom_feature_label() in flipped coordinates", {
   p <- base_flipped() + 
          geom_feature(data = example_features, 
-                      aes(x = position, y = molecule, forward = forward)) +
+                      aes(x = position, y = molecule, forward = forward),
+                      linewidth = 1) +
          geom_feature_label(data = example_features, 
                             aes(x = position, y = molecule, label = name, 
                                 forward = forward)) +
@@ -33,7 +35,8 @@ test_that("geom_feature() and geom_feature_label() in flipped coordinates", {
 test_that("geom_feature() and geom_feature_label() in polar coordinates", {
   p <- base_polar() + 
          geom_feature(data = example_features_polar, 
-                      aes(x = position, y = molecule, forward = forward)) +
+                      aes(x = position, y = molecule, forward = forward),
+                      linewidth = 1) +
          geom_feature_label(data = example_features_polar, 
                             aes(x = position, y = molecule, label = name, 
                                 forward = forward)) +
@@ -49,7 +52,8 @@ test_that("geom_feature() with numeric 'forward'", {
   ef2 <- example_features
   ef2$forward <- as.numeric(example_features$forward)
   p <- base_cartesian() +
-    geom_feature(data = ef2, aes(x = position, y = molecule, forward = forward)) +
+    geom_feature(data = ef2, aes(x = position, y = molecule, forward = forward),
+                 linewidth = 1) +
     geom_feature_label(data = ef2, aes(x = position, y = molecule, label = name, 
                                        forward = forward)) +
     geom_gene_arrow()